backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
apscheduler==3.10.4
psutil==6.0.0
pydantic==2.9.2
python-multipart==0.0.12
